[chore] Refactor app dependencies
