+++ fixbackend/app.py
@@ -16,15 +16,8 @@ import base64
 import logging
 import os
 from contextlib import asynccontextmanager
-from dataclasses import replace
-from datetime import timedelta
-from ssl import Purpose, create_default_context
-from typing import Any, AsyncIterator, Awaitable, Callable, ClassVar, Dict, Optional, Set, Tuple, cast
+from typing import Any, Awaitable, Callable, ClassVar, Optional, Set, Tuple, cast, AsyncIterator
 
-import boto3
-import httpx
-from arq import create_pool
-from arq.connections import RedisSettings
 from async_lru import alru_cache
 from fastapi import APIRouter, FastAPI, Request, Response
 from fastapi.exception_handlers import http_exception_handler
@@ -33,65 +26,31 @@ from fastapi.openapi.docs import get_swagger_ui_html
 from fastapi.responses import HTMLResponse, JSONResponse
 from fastapi.staticfiles import StaticFiles
 from fixcloudutils.logging import setup_logger
-from fixcloudutils.redis.event_stream import RedisStreamPublisher
-from fixcloudutils.redis.pub_sub import RedisPubSubPublisher
-from httpx import AsyncClient, Limits, Timeout
 from prometheus_fastapi_instrumentator import Instrumentator
-from redis.asyncio import Redis
 from sqlalchemy import select
-from sqlalchemy.ext.asyncio import async_sessionmaker, create_async_engine
 from starlette.exceptions import HTTPException
 
 from fixbackend import config, dependencies
-from fixbackend.analytics.domain_event_to_analytics import analytics
 from fixbackend.analytics.router import analytics_router
+from fixbackend.app_dependencies import create_dependencies
 from fixbackend.auth.auth_backend import cookie_transport
 from fixbackend.auth.depedencies import refreshed_session_scope
 from fixbackend.auth.oauth_router import github_client, google_client
 from fixbackend.auth.router import auth_router
-from fixbackend.auth.user_repository import UserRepository
 from fixbackend.auth.users_router import users_router
-from fixbackend.billing.billing_job import BillingJob
 from fixbackend.billing.router import billing_info_router
-from fixbackend.billing.service import BillingEntryService
-from fixbackend.certificates.cert_store import CertificateStore
-from fixbackend.cloud_accounts.account_setup import AwsAccountSetupHelper
-from fixbackend.cloud_accounts.repository import CloudAccountRepositoryImpl
 from fixbackend.cloud_accounts.router import cloud_accounts_callback_router, cloud_accounts_router
-from fixbackend.cloud_accounts.service_impl import CloudAccountServiceImpl
-from fixbackend.collect.collect_queue import RedisCollectQueue
 from fixbackend.config import Config
-from fixbackend.dependencies import FixDependencies
-from fixbackend.dependencies import ServiceNames as SN  # noqa
-from fixbackend.dispatcher.dispatcher_service import DispatcherService
-from fixbackend.dispatcher.next_run_repository import NextRunRepository
-from fixbackend.domain_events import DomainEventsStreamName
-from fixbackend.domain_events.consumers import EmailOnSignupConsumer
-from fixbackend.domain_events.publisher_impl import DomainEventPublisherImpl
-from fixbackend.domain_events.subscriber import DomainEventSubscriber
+from fixbackend.dependencies import ServiceNames as SN, FixDependency, FixDependencies  # noqa
 from fixbackend.errors import ClientError, NotAllowed, ResourceNotFound, WrongState
 from fixbackend.events.router import websocket_router
-from fixbackend.graph_db.service import GraphDatabaseAccessManager
-from fixbackend.inventory.inventory_client import InventoryClient, InventoryException
-from fixbackend.inventory.inventory_service import InventoryService
+from fixbackend.inventory.inventory_client import InventoryException
 from fixbackend.inventory.router import inventory_router
-from fixbackend.jwt import JwtServiceImpl
 from fixbackend.logging_context import get_logging_context, set_fix_cloud_account_id, set_workspace_id
-from fixbackend.metering.metering_repository import MeteringRepository
 from fixbackend.middleware.x_real_ip import RealIpMiddleware
-from fixbackend.notification.email.scheduled_email import ScheduledEmailSender
 from fixbackend.notification.notification_router import notification_router, unsubscribe_router
-from fixbackend.notification.notification_service import NotificationService
-from fixbackend.notification.user_notification_repo import UserNotificationSettingsRepositoryImpl
-from fixbackend.permissions.role_repository import RoleRepositoryImpl
 from fixbackend.permissions.router import roles_router
-from fixbackend.sqlalechemy_extensions import EngineMetrics
-from fixbackend.subscription.aws_marketplace import AwsMarketplaceHandler
 from fixbackend.subscription.router import subscription_router
-from fixbackend.subscription.stripe_subscription import create_stripe_service
-from fixbackend.subscription.subscription_repository import SubscriptionRepository
-from fixbackend.workspaces.invitation_repository import InvitationRepositoryImpl
-from fixbackend.workspaces.repository import WorkspaceRepositoryImpl
 from fixbackend.workspaces.router import workspaces_router
 
 log = logging.getLogger(__name__)
@@ -99,392 +58,25 @@ API_PREFIX = "/api"
 
 
 # noinspection PyUnresolvedReferences
-def fast_api_app(cfg: Config) -> FastAPI:
+async def fast_api_app(cfg: Config, deps: FixDependencies) -> FastAPI:
     google = google_client(cfg)
     github = github_client(cfg)
-    boto_session = boto3.Session(cfg.aws_access_key_id, cfg.aws_secret_access_key, region_name="us-east-1")
-    deps = FixDependencies()
-    deps.add(SN.config, cfg)
-    ca_cert_path = str(cfg.ca_cert) if cfg.ca_cert else None
-    client_context = create_default_context(purpose=Purpose.SERVER_AUTH)
-    if ca_cert_path:
-        client_context.load_verify_locations(ca_cert_path)
-    http_client = deps.add(
-        SN.http_client,
-        AsyncClient(
-            verify=client_context or True,
-            timeout=Timeout(pool=10, connect=10, read=60, write=60),
-            follow_redirects=True,
-            limits=Limits(max_connections=512, max_keepalive_connections=50),
-        ),
-    )
-    engine = deps.add(
-        SN.async_engine,
-        create_async_engine(cfg.database_url, pool_size=10, pool_recycle=3600, pool_pre_ping=True),
-    )
-    EngineMetrics.register(engine)
-    session_maker = deps.add(SN.session_maker, async_sessionmaker(engine))
 
-    def create_redis(url: str) -> Redis:
-        kwargs: Dict[str, Any] = dict(ssl_ca_certs=ca_cert_path) if url.startswith("rediss://") else {}
-        kwargs["health_check_interval"] = 30
-        kwargs["socket_timeout"] = 10
-        kwargs["socket_connect_timeout"] = 10
-        kwargs["socket_keepalive"] = True
-        if cfg.args.redis_password:
-            kwargs["password"] = cfg.args.redis_password
-        return Redis.from_url(url, decode_responses=True, **kwargs)  # type: ignore
+    @alru_cache(maxsize=1)
+    async def load_app_from_cdn() -> bytes:
+        app_url = f"{cfg.frontend_cdn_origin()}/index.html"
+        log.info(f"Loading app from CDN {app_url}")
+        response = await deps.http_client.get(app_url)
+        log.info("Loaded app from CDN")
+        body = response.content
+        return body
 
     @asynccontextmanager
-    async def setup_teardown_application(_: FastAPI) -> AsyncIterator[None]:
-        arq_redis = deps.add(
-            SN.arq_redis,
-            await create_pool(
-                replace(
-                    RedisSettings.from_dsn(cfg.redis_queue_url),
-                    ssl_ca_certs=ca_cert_path,
-                    password=cfg.args.redis_password,
-                )
-            ),
-        )
-        deps.add(SN.readonly_redis, create_redis(cfg.redis_readonly_url))
-        readwrite_redis = deps.add(SN.readwrite_redis, create_redis(cfg.redis_readwrite_url))
-        temp_store_redis = deps.add(SN.temp_store_redis, create_redis(cfg.redis_temp_store_url))
-        domain_event_subscriber = deps.add(
-            SN.domain_event_subscriber,
-            DomainEventSubscriber(readwrite_redis, cfg, "fixbackend"),
-        )
-        cloud_account_repo = deps.add(SN.cloud_account_repo, CloudAccountRepositoryImpl(session_maker))
-        deps.add(SN.next_run_repo, NextRunRepository(session_maker))
-        metering_repo = deps.add(SN.metering_repo, MeteringRepository(session_maker))
-        deps.add(SN.collect_queue, RedisCollectQueue(arq_redis))
-        graph_db_access = deps.add(SN.graph_db_access, GraphDatabaseAccessManager(cfg, session_maker))
-        inventory_client = deps.add(SN.inventory_client, InventoryClient(cfg.inventory_url, http_client))
-        inventory_service = deps.add(
-            SN.inventory,
-            InventoryService(
-                inventory_client, graph_db_access, cloud_account_repo, domain_event_subscriber, temp_store_redis
-            ),
-        )
-        fixbackend_events = deps.add(
-            SN.domain_event_redis_stream_publisher,
-            RedisStreamPublisher(
-                readwrite_redis,
-                DomainEventsStreamName,
-                "fixbackend",
-                keep_unprocessed_messages_for=timedelta(days=7),
-            ),
-        )
-        domain_event_publisher = deps.add(SN.domain_event_sender, DomainEventPublisherImpl(fixbackend_events))
-        subscription_repo = deps.add(SN.subscription_repo, SubscriptionRepository(session_maker))
-        user_repo = deps.add(SN.user_repo, UserRepository(session_maker))
-        role_repo = deps.add(SN.role_repository, RoleRepositoryImpl(session_maker))
-
-        workspace_repo = deps.add(
-            SN.workspace_repo,
-            WorkspaceRepositoryImpl(
-                session_maker,
-                graph_db_access,
-                domain_event_publisher,
-                RedisPubSubPublisher(
-                    redis=readwrite_redis,
-                    channel="workspaces",
-                    publisher_name="workspace_service",
-                ),
-                subscription_repo,
-                role_repo,
-            ),
-        )
-        billing_entry_service = deps.add(
-            SN.billing_entry_service,
-            BillingEntryService(
-                subscription_repo, workspace_repo, metering_repo, domain_event_publisher, cfg.billing_period
-            ),
-        )
-        analytics_event_sender = deps.add(
-            SN.analytics_event_sender, analytics(cfg, http_client, domain_event_subscriber, workspace_repo)
-        )
-        deps.add(
-            SN.invitation_repository,
-            InvitationRepositoryImpl(session_maker, workspace_repo, user_repository=user_repo),
-        )
-        deps.add(
-            SN.aws_marketplace_handler,
-            AwsMarketplaceHandler(
-                subscription_repo,
-                workspace_repo,
-                boto_session,
-                domain_event_publisher,
-                billing_entry_service,
-                cfg.args.aws_marketplace_metering_sqs_url,
-            ),
-        )
-        deps.add(
-            SN.stripe_service,
-            create_stripe_service(
-                cfg, user_repo, subscription_repo, workspace_repo, session_maker, domain_event_publisher
-            ),
-        )
-        cloud_accounts_redis_publisher = RedisPubSubPublisher(
-            redis=readwrite_redis,
-            channel="cloud_accounts",
-            publisher_name="cloud_account_service",
-        )
-
-        cert_store = deps.add(SN.certificate_store, CertificateStore(cfg))
-
-        jwt_service = deps.add(SN.jwt_service, JwtServiceImpl(cert_store))
-
-        notification_service = deps.add(
-            SN.notification_service,
-            NotificationService(
-                cfg,
-                workspace_repo,
-                graph_db_access,
-                user_repo,
-                inventory_service,
-                readwrite_redis,
-                session_maker,
-                http_client,
-                domain_event_publisher,
-                domain_event_subscriber,
-                jwt_service,
-            ),
-        )
-        deps.add(SN.email_on_signup_consumer, EmailOnSignupConsumer(notification_service, domain_event_subscriber))
-        deps.add(
-            SN.cloud_account_service,
-            CloudAccountServiceImpl(
-                workspace_repository=workspace_repo,
-                cloud_account_repository=CloudAccountRepositoryImpl(session_maker),
-                pubsub_publisher=cloud_accounts_redis_publisher,
-                domain_event_publisher=domain_event_publisher,
-                readwrite_redis=readwrite_redis,
-                config=cfg,
-                account_setup_helper=AwsAccountSetupHelper(boto_session),
-                dispatching=False,
-                http_client=http_client,
-                boto_session=boto_session,
-                cf_stack_queue_url=cfg.aws_cf_stack_notification_sqs_url,
-                notification_service=notification_service,
-                analytics_event_sender=analytics_event_sender,
-            ),
-        )
-        deps.add(SN.user_notification_settings_repository, UserNotificationSettingsRepositoryImpl(session_maker))
-        if not cfg.static_assets:
-            await load_app_from_cdn()
-        async with deps:
-            log.info("Application services started.")
-            yield None
-        await arq_redis.close()
-        log.info("Application services stopped.")
-
-    @asynccontextmanager
-    async def setup_teardown_dispatcher(_: FastAPI) -> AsyncIterator[None]:
-        arq_redis = deps.add(
-            SN.arq_redis,
-            await create_pool(
-                replace(
-                    RedisSettings.from_dsn(cfg.redis_queue_url),
-                    ssl_ca_certs=ca_cert_path,
-                    password=cfg.args.redis_password,
-                )
-            ),
-        )
-        readwrite_redis = deps.add(SN.readwrite_redis, create_redis(cfg.redis_readwrite_url))
-        domain_event_subscriber = deps.add(
-            SN.domain_event_subscriber,
-            DomainEventSubscriber(readwrite_redis, cfg, "dispatching"),
-        )
-        temp_store_redis = deps.add(SN.temp_store_redis, create_redis(cfg.redis_temp_store_url))
-        cloud_account_repo = deps.add(SN.cloud_account_repo, CloudAccountRepositoryImpl(session_maker))
-        next_run_repo = deps.add(SN.next_run_repo, NextRunRepository(session_maker))
-        metering_repo = deps.add(SN.metering_repo, MeteringRepository(session_maker))
-        collect_queue = deps.add(SN.collect_queue, RedisCollectQueue(arq_redis))
-        graph_db_access = deps.add(SN.graph_db_access, GraphDatabaseAccessManager(cfg, session_maker))
-        fixbackend_events = deps.add(
-            SN.domain_event_redis_stream_publisher,
-            RedisStreamPublisher(
-                readwrite_redis,
-                DomainEventsStreamName,
-                "dispatching",
-                keep_unprocessed_messages_for=timedelta(days=7),
-            ),
-        )
-        domain_event_publisher = deps.add(SN.domain_event_sender, DomainEventPublisherImpl(fixbackend_events))
-        subscription_repo = deps.add(SN.subscription_repo, SubscriptionRepository(session_maker))
-        role_repo = deps.add(SN.role_repository, RoleRepositoryImpl(session_maker))
-
-        workspace_repo = deps.add(
-            SN.workspace_repo,
-            WorkspaceRepositoryImpl(
-                session_maker,
-                graph_db_access,
-                domain_event_publisher,
-                RedisPubSubPublisher(
-                    redis=readwrite_redis,
-                    channel="workspaces",
-                    publisher_name="workspace_service",
-                ),
-                subscription_repo,
-                role_repo,
-            ),
-        )
-
-        # uncomment once aws marketplace suscriptions are available on prd
-        # trial_end_service = deps.add(SN.trial_end_service, TrialEndService(workspace_repo))
-
-        cloud_accounts_redis_publisher = RedisPubSubPublisher(
-            redis=readwrite_redis,
-            channel="cloud_accounts",
-            publisher_name="cloud_account_service",
-        )
-        user_repo = deps.add(SN.user_repo, UserRepository(session_maker))
-        inventory_client = deps.add(SN.inventory_client, InventoryClient(cfg.inventory_url, http_client))
-        # in dispatching we do not want to handle domain events: leave it to the app
-        inventory_service = deps.add(
-            SN.inventory,
-            InventoryService(inventory_client, graph_db_access, cloud_account_repo, None, temp_store_redis),
-        )
-
-        cert_store = deps.add(SN.certificate_store, CertificateStore(cfg))
-
-        jwt_service = deps.add(SN.jwt_service, JwtServiceImpl(cert_store))
-
-        notification_service = deps.add(
-            SN.notification_service,
-            NotificationService(
-                cfg,
-                workspace_repo,
-                graph_db_access,
-                user_repo,
-                inventory_service,
-                readwrite_redis,
-                session_maker,
-                http_client,
-                domain_event_publisher,
-                domain_event_subscriber,
-                jwt_service,
-                handle_events=False,  # fixbackend will handle events. dispatching should ignore them.
-            ),
-        )
-        analytics_event_sender = deps.add(
-            SN.analytics_event_sender, analytics(cfg, http_client, domain_event_subscriber, workspace_repo)
-        )
-        deps.add(
-            SN.cloud_account_service,
-            CloudAccountServiceImpl(
-                workspace_repository=workspace_repo,
-                cloud_account_repository=CloudAccountRepositoryImpl(session_maker),
-                pubsub_publisher=cloud_accounts_redis_publisher,
-                domain_event_publisher=domain_event_publisher,
-                readwrite_redis=readwrite_redis,
-                config=cfg,
-                account_setup_helper=AwsAccountSetupHelper(boto_session),
-                dispatching=True,
-                http_client=http_client,
-                boto_session=boto_session,
-                cf_stack_queue_url=cfg.aws_cf_stack_notification_sqs_url,
-                notification_service=notification_service,
-                analytics_event_sender=analytics_event_sender,
-            ),
-        )
-        deps.add(
-            SN.dispatching,
-            DispatcherService(
-                readwrite_redis,
-                cloud_account_repo,
-                next_run_repo,
-                metering_repo,
-                collect_queue,
-                graph_db_access,
-                domain_event_publisher,
-                temp_store_redis,
-                domain_event_subscriber,
-                workspace_repo,
-            ),
-        )
-        deps.add(SN.scheduled_email_sender, ScheduledEmailSender(notification_service.email_sender, session_maker))
-
-        async with deps:
-            log.info("Application services started.")
-            yield None
-        await arq_redis.close()
-        log.info("Application services stopped.")
-
-    @asynccontextmanager
-    async def setup_teardown_billing(_: FastAPI) -> AsyncIterator[None]:
-        graph_db_access = deps.add(SN.graph_db_access, GraphDatabaseAccessManager(cfg, session_maker))
-        readwrite_redis = deps.add(SN.readwrite_redis, create_redis(cfg.redis_readwrite_url))
-        fixbackend_events = deps.add(
-            SN.domain_event_redis_stream_publisher,
-            RedisStreamPublisher(
-                readwrite_redis,
-                DomainEventsStreamName,
-                "fixbackend",
-                keep_unprocessed_messages_for=timedelta(days=7),
-            ),
-        )
-        domain_event_publisher = deps.add(SN.domain_event_sender, DomainEventPublisherImpl(fixbackend_events))
-        metering_repo = deps.add(SN.metering_repo, MeteringRepository(session_maker))
-        subscription_repo = deps.add(SN.subscription_repo, SubscriptionRepository(session_maker))
-        role_repo = deps.add(SN.role_repository, RoleRepositoryImpl(session_maker))
-        user_repo = deps.add(SN.user_repo, UserRepository(session_maker))
-        workspace_repo = deps.add(
-            SN.workspace_repo,
-            WorkspaceRepositoryImpl(
-                session_maker,
-                graph_db_access,
-                domain_event_publisher,
-                RedisPubSubPublisher(
-                    redis=readwrite_redis,
-                    channel="workspaces",
-                    publisher_name="workspace_service",
-                ),
-                subscription_repo,
-                role_repo,
-            ),
-        )
-        billing_entry_service = deps.add(
-            SN.billing_entry_service,
-            BillingEntryService(
-                subscription_repo, workspace_repo, metering_repo, domain_event_publisher, cfg.billing_period
-            ),
-        )
-        aws_marketplace = deps.add(
-            SN.aws_marketplace_handler,
-            AwsMarketplaceHandler(
-                subscription_repo,
-                workspace_repo,
-                boto_session,
-                domain_event_publisher,
-                billing_entry_service,
-                cfg.args.aws_marketplace_metering_sqs_url,
-            ),
-        )
-        stripe = deps.add(
-            SN.stripe_service,
-            create_stripe_service(
-                cfg, user_repo, subscription_repo, workspace_repo, session_maker, domain_event_publisher
-            ),
-        )
-        deps.add(
-            SN.billing_job,
-            BillingJob(aws_marketplace, stripe, subscription_repo, workspace_repo, billing_entry_service, cfg),
-        )
-
+    async def lifespan(_: FastAPI) -> AsyncIterator[None]:
         async with deps:
             log.info("Application services started.")
             yield None
         log.info("Application services stopped.")
-
-    match cfg.args.mode:
-        case "dispatcher":
-            lifespan = setup_teardown_dispatcher
-        case "billing":
-            lifespan = setup_teardown_billing
-        case _:
-            lifespan = setup_teardown_application
 
     app = FastAPI(
         title="Fix Backend",
@@ -568,16 +160,6 @@ def fast_api_app(cfg: Config) -> FastAPI:
 
     # Add filter to the logger
     logging.getLogger("uvicorn.access").addFilter(EndpointFilter())
-
-    @alru_cache(maxsize=1)
-    async def load_app_from_cdn() -> bytes:
-        async with httpx.AsyncClient() as client:
-            app_url = f"{cfg.frontend_cdn_origin()}/index.html"
-            log.info(f"Loading app from CDN {app_url}")
-            response = await client.get(app_url)
-            log.info("Loaded app from CDN")
-            body = response.content
-            return body
 
     @app.get("/health", tags=["system"])
     async def health() -> Response:
@@ -708,7 +290,7 @@ def fast_api_app(cfg: Config) -> FastAPI:
     return app
 
 
-def setup_process() -> FastAPI:
+async def setup_process() -> FastAPI:
     """
     This function is used by uvicorn to start the server.
     Entrypoint for the application to start the server.
@@ -721,10 +303,5 @@ def setup_process() -> FastAPI:
         get_logging_context=get_logging_context,
     )
 
-    # Replace all special uvicorn handlers
-    for logger in ["uvicorn", "uvicorn.error", "uvicorn.access"]:
-        lg = logging.getLogger(logger)
-        lg.handlers.clear()  # remove handlers
-        lg.propagate = True  # propagate to root, so the handlers there are used
-
-    return fast_api_app(current_config)
+    deps = await create_dependencies(current_config)
+    return await fast_api_app(current_config, deps)
